+++ src/frontend/src/components/login/ConnectButton.tsx
@@ -1,16 +1,17 @@
 import Button from "../ui/Button";
+import ConnectDialog from "../ConnectDialog";
 import { faCircleNotch } from "@fortawesome/free-solid-svg-icons";
 import { faEthereum } from "@fortawesome/free-brands-svg-icons";
 import { useAccount } from "wagmi";
-import { useConnectModal } from "@rainbow-me/rainbowkit";
+import { useState } from "react";
 
 export default function ConnectButton() {
   const { isConnecting } = useAccount();
-
-  const { openConnectModal } = useConnectModal();
+  const [connectDialogOpen, setConnectDialogOpen] = useState(false);
 
   const handleClick = () => {
-    openConnectModal?.();
+    if (isConnecting) return;
+    setConnectDialogOpen(true);
   };
 
   const buttonIcon = isConnecting ? faCircleNotch : faEthereum;
@@ -28,6 +29,10 @@ export default function ConnectButton() {
       >
         {buttonText}
       </Button>
+      <ConnectDialog
+        isOpen={connectDialogOpen}
+        setIsOpen={setConnectDialogOpen}
+      />
     </>
   );
 }