+++ src/frontend/src/eth/utils/shortenEthAddress.ts
@@ -1,0 +1,14 @@
+/**
+ * Shorten ethereum address to a string formatted as:
+ *  0x, first 4 digits, ellipsis, last 4 digits
+ *
+ * @exports
+ * @param {string} address
+ * @returns {string}
+ */
+export const shortenEthAddress = (address?: string): string | undefined => {
+  if (!address) return address;
+  return `${address.substring(0, 6)}...${address.substring(
+    address.length - 4
+  )}`;
+};