+++ src/frontend/src/main.tsx
@@ -1,12 +1,10 @@
 import "./index.css";
-import "@rainbow-me/rainbowkit/styles.css";
 
-import { RainbowKitProvider, darkTheme } from "@rainbow-me/rainbowkit";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import {
   canisterId,
   idlFactory,
 } from "../../declarations/ic_siwe_provider/index";
-import { chains, wagmiConfig } from "./wagmi/wagmi.config.ts";
 
 import Actors from "./ic/Actors.tsx";
 import App from "./App.tsx";
@@ -15,22 +13,16 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { SiweIdentityProvider } from "ic-use-siwe-identity";
 import { Toaster } from "react-hot-toast";
-import { WagmiConfig } from "wagmi";
+import { WagmiProvider } from "wagmi";
 import { _SERVICE } from "../../declarations/ic_siwe_provider/ic_siwe_provider.did";
+import { wagmiConfig } from "./wagmi/wagmi.config.ts";
+
+const queryClient = new QueryClient();
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <WagmiConfig config={wagmiConfig}>
-      <RainbowKitProvider
-        chains={chains}
-        modalSize="compact"
-        theme={darkTheme({
-          accentColor: "#7b3fe4",
-          accentColorForeground: "white",
-          borderRadius: "large",
-          overlayBlur: "none",
-        })}
-      >
+    <WagmiProvider config={wagmiConfig}>
+      <QueryClientProvider client={queryClient}>
         <SiweIdentityProvider<_SERVICE>
           canisterId={canisterId}
           idlFactory={idlFactory}
@@ -41,8 +33,8 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
             </AuthGuard>
           </Actors>
         </SiweIdentityProvider>
-      </RainbowKitProvider>
-    </WagmiConfig>
+      </QueryClientProvider>
+    </WagmiProvider>
     <Toaster />
   </React.StrictMode>
 );