+++ src/frontend/src/wagmi/is-chain-id-supported.tsx
@@ -1,5 +1,5 @@
-import { supportedChains } from "./wagmi.config";
+import { wagmiConfig } from "./wagmi.config";
 
 export function isChainIdSupported(id?: number) {
-  return supportedChains.find((c) => c.id === id) !== undefined;
+  return wagmiConfig.chains.find((c) => c.id === id) !== undefined;
 }